SE2019/requirements.txt
aioredis==1.2.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.2
asgiref==3.1.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.2
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
clyent==1.2.2
colorama==0.4.1
conda==4.6.14
conda-build==3.17.8
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
daphne==2.3.0
dask==1.2.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.5.0
distributed==1.26.1
Django==2.1.7
django-appconf==1.0.3
django-crontab==0.7.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-widget-tweaks==1.4.3
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
h5py==2.9.0
heapdict==1.0.0
hiredis==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.0.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jeepney==0.4
jieba==0.39
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
meld3==1.0.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mysqlclient==1.3.14
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
olefile==0.46
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
path.py==12.0.1
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
pilkit==2.0
Pillow==6.0.0
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
requests==2.21.0
rope==0.12.0
ruamel-yaml==0.15.46
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
sphinxcontrib-websupport==1.1.0
spyder==3.3.4
spyder-kernels==0.4.4
SQLAlchemy==1.3.1
statsmodels==0.9.0
supervisor==4.0.2
sympy==1.3
tables==3.5.1
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
Twisted==19.2.0
txaio==18.8.1
unicodecsv==0.14.1
urllib3==1.24.1
uWSGI==2.0.18
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
Whoosh==2.7.4
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.6
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
zope.interface==4.6.0
